requirements.txt
networkx==3.2.1
numpy==1.26.3
torch==1.13.1
torch_geometric==2.6.1
matplotlib==3.7.0
scipy==1.10.1
scikit-learn==1.5.1
umap-learn==0.5.6
pandas==2.2.0
notebook==6.5.6
pyarrow==11.0.0
ipywidgets==8.0.4
